requirements.txt
langchain 
langchain-community 
langchain-openai 
langchain-experimental 
tiktoken  
python-dotenv
fastapi
fastapi[standard] 
uvicorn
python-multipart
pydantic
requests  
langchain-community
pytest
langchain_chroma
langgraph
llama_index
langchain_mongodb
langchain-text-splitters
forex_python
pandas_ta
yfinance
chromadb 
pypdf 
openai 
tiktoken